+++ src/cas.rs
@@ -1,5 +1,35 @@
 use super::codec::SupportedCodecs;
+use anyhow::Result;
 use libipld::cid::Cid;
+use rocket::{
+    form::{DataField, FromFormField},
+    request::FromParam,
+};
+use std::str::FromStr;
+
+#[derive(PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct CidWrap(pub Cid);
+
+// Orphan rule requires a wrapper type for this :(
+impl<'a> FromParam<'a> for CidWrap {
+    type Error = anyhow::Error;
+    fn from_param(param: &'a str) -> Result<CidWrap> {
+        Ok(CidWrap(Cid::from_str(param)?))
+    }
+}
+
+#[rocket::async_trait]
+impl<'r> FromFormField<'r> for CidWrap {
+    async fn from_data(field: DataField<'r, '_>) -> rocket::form::Result<'r, Self> {
+        Ok(CidWrap(
+            field
+                .name
+                .source()
+                .parse()
+                .map_err(|_| field.unexpected())?,
+        ))
+    }
+}
 
 #[rocket::async_trait]
 pub trait ContentAddressedStorage: Send + Sync {
